Move bvh and aabb to mod scene::object
