--- src/scene/bvh/mod.rs
@@ -1,5 +1,0 @@
-mod aabb;
-mod bvh;
-
-pub use aabb::AABB;
-pub use bvh::*;
