+++ src/scene/hit.rs
@@ -4,7 +4,7 @@ use crate::materials::Material;
 use super::hitrecord::BasicHitRecord;
 
 use crate::core::time::{Interval, TimeRay3};
-use crate::scene::bvh::AABB;
+use crate::scene::object::AABB;
 use std::cmp::Ordering;
 use std::rc::Rc;
