+++ src/scene/mod.rs
@@ -1,4 +1,3 @@
-pub mod bvh;
 pub mod camera;
 mod hit;
 mod hitrecord;
