+++ src/scene/object/mod.rs
@@ -1,2 +1,7 @@
 pub mod movingsphere;
 pub mod sphere;
+mod aabb;
+mod bvh;
+
+pub use aabb::*;
+pub use bvh::*;
